More debug

+++ db/migrate/20231121195314_drop_requested_item.rb
@@ -1,5 +1,5 @@
 class DropRequestedItem < ActiveRecord::Migration[6.1]
   def change
-    drop_table :food_requests
+    drop_table :food_requests if ActiveRecord::Base.connection.table_exists?('food_requests')
   end
 end
